+++ apps/payments/app/[locale]/(hosted)/paymentSetup/PaymentSetupForm.tsx
@@ -1,7 +1,11 @@
 import { getTranslations } from "next-intl/server";
 import { PaymentRequestDetails } from "./db";
-import { providerTypes } from "./providers/types";
 import buildApiClient from "../../../../client/index";
+import {
+  paymentMethodToProviderType,
+  paymentMethods,
+  providerTypeToPaymentMethod,
+} from "../../../utils";
 
 async function getRegisteredAccounts(userId: string) {
   const providers = (await buildApiClient(userId).providers.apiV1ProvidersGet())
@@ -12,11 +16,12 @@ async function getRegisteredAccounts(userId: string) {
   }
 
   const accounts = providers.reduce((acc, provider) => {
-    if (!acc.get(provider.type)) {
-      acc.set(provider.type, []);
+    const paymentMethod = providerTypeToPaymentMethod[provider.type];
+    if (!acc.get(paymentMethod)) {
+      acc.set(paymentMethod, []);
     }
 
-    acc.get(provider.type).push({
+    acc.get(paymentMethod).push({
       id: provider.id,
       name: provider.name,
     });
@@ -72,30 +77,28 @@ export default async function ({
             defaultValue={details?.description}
           ></textarea>
         </div>
-        {providerTypes.map((providerType, index) => {
-          // TODO: remove this line once worldpay is integrated
-          if (providerType === "worldpay") return null;
-          const provider = details?.providers.find(
-            (p) => p.type === providerType,
+        {paymentMethods.map((paymentMethod, index) => {
+          const provider = details?.providers.find((p) =>
+            paymentMethodToProviderType[paymentMethod].includes(p.type),
           );
           return (
             <div className="govie-form-group" key={index}>
               <label
-                htmlFor={`${providerType}-account`}
+                htmlFor={`${paymentMethod}-account`}
                 className="govie-label--s"
               >
-                {t(`form.paymentProvider.${providerType}`)}
+                {t(`form.paymentProvider.${paymentMethod}`)}
               </label>
               <br />
               <select
-                id={`${providerType}-account`}
-                name={`${providerType}-account`}
+                id={`${paymentMethod}-account`}
+                name={`${paymentMethod}-account`}
                 className="govie-select"
                 defaultValue={provider?.id}
                 style={{ width: "350px" }}
               >
                 <option value={""}>Disabled</option>
-                {(providerAccounts.get(providerType) ?? []).map((account) => (
+                {(providerAccounts.get(paymentMethod) ?? []).map((account) => (
                   <option key={account.id} value={account.id}>
                     {account.name}
                   </option>
