+++ apps/payments/app/[locale]/(hosted)/paymentSetup/create/page.tsx
@@ -7,12 +7,10 @@ import buildApiClient from "../../../../../client/index";
 async function createPayment(userId: string, formData: FormData) {
   "use server";
 
-  // Worldpay integration is missing, temporarily ignore it
   const providers: string[] = [
     formData.get("openbanking-account")?.toString(),
     formData.get("banktransfer-account")?.toString(),
-    formData.get("stripe-account")?.toString(),
-    formData.get("realex-account")?.toString(),
+    formData.get("card-account")?.toString(),
   ].filter((provider): provider is string => !!provider);
 
   const data = {
