+++ apps/payments/app/[locale]/(hosted)/paymentSetup/db.ts
@@ -1,4 +1,5 @@
 import { PaymentRequestDO } from "../../../../types/common";
+import { ProviderType } from "./providers/types";
 
 export type PaymentRequestDetails = Pick<
   PaymentRequestDO,
@@ -13,7 +14,7 @@ export type PaymentRequestDetails = Pick<
 > & {
   providers: {
     name: string;
-    type: string;
+    type: ProviderType;
     id: string;
   }[];
 };
